move package directory
