+++ internal/frontend/handlers/dag.go
@@ -1,4 +1,4 @@
-package dag
+package handlers
 
 import (
 	"context"
